Move transformations to transform.js
