+++ packages/edit-navigation/src/store/test/transform.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import menuItemsToBlocks from '../menu-items-to-blocks';
+import { menuItemsToBlocks } from '../transform';
 
 // Mock createBlock to avoid creating the blocks in test environment.
 jest.mock( '@wordpress/blocks', () => {
